+++ tests/test-rounding.R
@@ -3,10 +3,7 @@ library( munsellinterpol )
 
 options( width=144 )
 
-#
-#   checkColorLookup()
-#   performs lookup with ColorBlockFromMunsell() and verifies that the color block number is correct
-#
+
 #   return value:  TRUE or FALSE
 #
 testRounding <- function( count=100 )